Added Constabulary
Is the patrol base but fleet cap changes depending on market stability.
Specifics:
Calculates a ratio of current market stability / 9 ( a high stability)
the lower the stability is, the more light and medium patrols the planet gets.
However if the planet has > 9 stability, give a bonus of 2 heavy fleets instead.